Fixed incorrect price helper usage
no refs

The new price helper expects a plan object or amount which is formatted in the currency format, but on the signup page here we don't have access to plan object, and it does not replace the monthly/yearly plan values being returned by `@price.monthly` or `@price.yearly`.

The fix updates the usage of `price` helper to use the amount from `@price` data and add currency symbol using new price helper update which allows currency symbol output from the currency value.